*** empty log message ***
git-svn-id: https://hpop.svn.sourceforge.net/svnroot/hpop/trunk@8 0ca4f280-b5d2-4e9a-a707-da1077cca493

+++ hpop/TNEF/TNEFParser.cs
@@ -336,7 +336,7 @@ namespace OpenPOP.TNEF
 
 					byte[] _subjectBuffer=new byte[len-1];
 
-					Array.Copy(buf,_subjectBuffer,(long)len-1);
+					Array.Copy(buf,_subjectBuffer,(int)len-1);
 
 					strSubject=Encoding.Default.GetString(_subjectBuffer);
 
@@ -351,7 +351,7 @@ namespace OpenPOP.TNEF
 					StreamReadBytes(buf,len);
 					//PrintResult("File-Name: {0}\n", buf);
 					byte[] _fileNameBuffer=new byte[len-1];
-					Array.Copy(buf,_fileNameBuffer,(long)len-1);
+					Array.Copy(buf,_fileNameBuffer,(int)len-1);
 
 					if (_fileNameBuffer == null) _fileNameBuffer = Encoding.Default.GetBytes("tnef.dat");
 					string strFileName=Encoding.Default.GetString(_fileNameBuffer);
